requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
certifi==2021.5.30
charset-normalizer==2.0.4
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
django-cors-headers==3.9.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
gunicorn==20.1.0
idna==3.2
mysql-connector-python==8.0.27
mysqlclient==2.0.3
platformdirs==2.2.0
protobuf==3.19.1
psycopg2==2.9.1
PyJWT==2.1.0
pytz==2021.1
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.6
virtualenv==20.7.0
whitenoise==5.3.0
